Moved context._amulet_app to _app
This package used to be more complex with fbs integration but it no longer needs to be a package.
