+++ src/amulet_editor/application/_app.py
